third_party/blink/web_tests/editing/execCommand/queryCommandSupported.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_false(document.queryCommandSupported('pasteFromImageURL'), '1 pasteFromImageURL');
}, 'queryCommandSupported for PasteFromImageURL');
</script>
